requirements.txt
openai==1.3.0
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
tqdm==4.66.1
pandas==2.1.3 